requirements.txt
fastapi
uvicorn
google-cloud-vision
pdf2image
pillow
requests
